+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Expression/Case.hs
@@ -1,0 +1,8 @@
+module Database.Persist.Sql.Lifted.Expression.Case
+  ( case_
+  , when_
+  , then_
+  , else_
+  ) where
+
+import Database.Esqueleto.Experimental