+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Expression/Comparison.hs
@@ -1,0 +1,14 @@
+module Database.Persist.Sql.Lifted.Expression.Comparison
+  ( -- * Equality
+    (==.)
+  , (!=.)
+
+    -- * Less & greater
+  , (>=.)
+  , (>.)
+  , (<=.)
+  , (<.)
+  , between
+  ) where
+
+import Database.Esqueleto.Experimental