+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Expression/Exists.hs
@@ -1,0 +1,6 @@
+module Database.Persist.Sql.Lifted.Expression.Exists
+  ( exists
+  , notExists
+  ) where
+
+import Database.Esqueleto.Experimental