+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Expression/List.hs
@@ -1,0 +1,9 @@
+module Database.Persist.Sql.Lifted.Expression.List
+  ( in_
+  , notIn
+  , subList_select
+  , valList
+  , justList
+  ) where
+
+import Database.Esqueleto.Experimental