+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Expression/Maybe.hs
@@ -1,0 +1,11 @@
+module Database.Persist.Sql.Lifted.Expression.Maybe
+  ( isNothing
+  , isNothing_
+  , just
+  , nothing
+  , joinV
+  , coalesce
+  , coalesceDefault
+  ) where
+
+import Database.Esqueleto.Experimental