+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Expression/String.hs
@@ -1,0 +1,18 @@
+module Database.Persist.Sql.Lifted.Expression.String
+  ( lower_
+  , upper_
+  , trim_
+  , ltrim_
+  , rtrim_
+  , length_
+  , left_
+  , right_
+  , like
+  , ilike
+  , (%)
+  , concat_
+  , (++.)
+  , castString
+  ) where
+
+import Database.Esqueleto.Experimental