+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Expression/SubSelect.hs
@@ -1,0 +1,10 @@
+module Database.Persist.Sql.Lifted.Expression.SubSelect
+  ( subSelect
+  , subSelectMaybe
+  , subSelectCount
+  , subSelectForeign
+  , subSelectList
+  , subSelectUnsafe
+  ) where
+
+import Database.Esqueleto.Experimental