+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Expression/Table.hs
@@ -1,0 +1,6 @@
+module Database.Persist.Sql.Lifted.Expression.Table
+  ( getTable
+  , getTableMaybe
+  ) where
+
+import Database.Esqueleto.Experimental