+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Query.hs
@@ -1,0 +1,50 @@
+module Database.Persist.Sql.Lifted.Query
+  ( -- * Type
+    SqlQuery
+
+    -- * Where
+  , where_
+
+    -- * Aggregate
+  , groupBy
+  , groupBy_
+  , having
+
+    -- * Limit & offset
+  , limit
+  , offset
+
+    -- * Distinct & order by
+  , distinct
+  , distinctOn
+  , orderBy
+  , don
+  , distinctOnOrderBy
+
+    -- * Update
+  , set
+
+    -- * withNonNull
+  , withNonNull
+
+    -- * Locking
+  , locking
+  , LockingKind (..)
+
+    -- * Set operations
+  , union_
+  , unionAll_
+  , except_
+  , intersect_
+
+    -- * Common table expressions
+  , with
+  , withRecursive
+  ) where
+
+import Database.Persist.Sql.Lifted.Query.Aggregate
+import Database.Persist.Sql.Lifted.Query.CommonTableExpressions
+import Database.Persist.Sql.Lifted.Query.Core
+import Database.Persist.Sql.Lifted.Query.Locking
+import Database.Persist.Sql.Lifted.Query.SetOperations
+import Database.Persist.Sql.Lifted.Query.Update