+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Query/Aggregate.hs
@@ -1,0 +1,7 @@
+module Database.Persist.Sql.Lifted.Query.Aggregate
+  ( groupBy
+  , groupBy_
+  , having
+  ) where
+
+import Database.Esqueleto.Experimental