+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Query/CommonTableExpressions.hs
@@ -1,0 +1,6 @@
+module Database.Persist.Sql.Lifted.Query.CommonTableExpressions
+  ( with
+  , withRecursive
+  ) where
+
+import Database.Esqueleto.Experimental