+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Query/SetOperations.hs
@@ -1,0 +1,9 @@
+module Database.Persist.Sql.Lifted.Query.SetOperations
+  ( union_
+  , unionAll_
+  , except_
+  , intersect_
+  )
+where
+
+import Database.Esqueleto.Experimental