+++ persistent-sql-lifted/library/Database/Persist/Sql/Lifted/Query/Update.hs
@@ -1,0 +1,5 @@
+module Database.Persist.Sql.Lifted.Query.Update
+  ( set
+  ) where
+
+import Database.Esqueleto.Experimental